functions/requirements.txt
firebase-functions>=0.4.0
firebase-admin>=6.2.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.4.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
mangum>=0.17.0
